move PV validator in separate package
